requirements.txt
pydantic
google-generativeai
openpyxl
fastapi 
uvicorn 
pandas